More general stat improvements, change stats to be doubles by default

+++ core/src/com/pottda/game/model/Inventory.java
@@ -70,7 +70,7 @@ public class Inventory {
 
         double sum = 0;
         for (Item i : items) {
-            sum += i.getStat(stat).doubleValue();
+            sum += i.getStat(stat);
         }
         return sum;
     }
